Reorganized lib/libd functions and Lavalink
