+++ src/commands/info.ts
@@ -2,7 +2,8 @@ import {MessageEmbed, version as djsversion} from "discord.js";
 import ms from "ms";
 import os from "os";
 import Command from "../core/command";
-import {formatBytes, trimArray, getMemberByUsername} from "../core/libd";
+import {formatBytes, trimArray} from "../core/lib";
+import {getMemberByUsername} from "../core/libd";
 import {verificationLevels, filterLevels, regions} from "../defs/info";
 import moment from "moment";
 import utc from "moment";
